Allow out-of-bound fence requests
The current behaviour is to throw an error if any coordinate provided in
the request falls outside the boundaries of the seismic volume. However,
with this pull request, the user will now be able to submit requests
containing such coordinates and the response will return fillvalues for
all values lying outside the seismic cube.