Add align_surfaces
We are planning to give user ability to calculate data between two
surfaces. Yet it could happen that surfaces are not aligned. Thus
alignment would always be a first step in the process.

Additional operation, like finding top/bottom surface, is done in
the same loop to limit the number of times we have to iterate over the
surface.