Add sub-slicing/crop support to core slice function
This commit extends the core API to accept a set of axis bounds which
are used crop the requested slice. This feature is useful both
internally in the code base, as well as providing it through the public
API.

Internally it can be used to distribute single requests a cross multiple
OpenVDS handles, or even across multiple nodes if we ever go fully
distributed.

As far as the users are concerned requesting subslices will result in
faster response times and thus provide an optimization over cropping at
client side - provided that's what they are interested in of course.
This is especially useful for large volumes.