Extract plane boundary logic out of RegularSurface
RegularSurface has been a bit confusing for me for a long time.

In theory RegularSurace represents a simple array, when in practice
class has more logic inside of it.
Each cell contains some data, but it also represents some CDP
coordinate. Thus it can be confusing - what is the real important value
of the cell, and what is just secondary helpful information? Will
surface[i] return me the stored data or cell's CDP?

Hence comes the idea to separate data and everything that is related to
the horizontal plane calculations. To do this Plane class is extended to
BoundedPlane which knows about its boundary limitations and thus can
take onto all the responsibilities that deal with coordinates
conversion.

Additionally it should give a benefit of clarity for some code.
In attributes calculation we have 3 Regular Surfaces which differ only
in its data, but which planes along with rows/columns must be the same.
Having BoundedPlane field makes this fact more obvious.