Prepare metadata for testing on various files
The ISO 8601 time format allows for optional inclusion of the "Z" time
zone offset designator at the end, indicating that the time is in UTC,
but the offset may also be specified in other ways depending on the
specifics of the time representation.

Also, added regex for the vds filename to accommodate different vds
files in the same testcase.