Get axis order from vds file
Initially, we assumed that the inline, crossline, and sample dimensions
in the VDS file would always be in the same order. However, this is not
always the case, and our code may throw an exception. To address this
issue, these changes will enable us to retrieve the inline, crossline,
and sample axis directly from the VDS file, even if they are ordered
differently in the file.