Extract common logic from data endpoint requests
We can notice that all data endpoints actually follow the same logic,
which we now copy again and again.
We make connection, check if the result is cached, and if it is not, we
calculate data and metadata.
Thus this logic can be extracted.

metadata endpoint is different and doesn't fit in, as it returns only
metadata part and we don't see feasible for it to be cached.