Introduce Segment
Segment is supposed to represent each vertical part of the 3D data in
attributes calculations.

It is supposed to replace AttributeComputeParams (hopefully it should be
enough to provide all the information needed for calculation of
attributes on the segment itself) and encapsulate vectors created on
resampling.

The was an idea before to introduce this class, but it was thrown away
as not feasible - creating 10 mln objects for just a couple of
operations.
However we anyway have to create different objects - for example vectors
that will contain new data on resampling. So introduction of this
conceptually useful class was warranted.

To improve performance object reinitialization is introduced.
On manual tests it appeared that object reinitialization instead of new
object creation gives attributes function up to 20% speed up. Real speed
up would depend on many factors.

With this approach overall performance of the solution doesn't seem to
be much different to the one before refactoring.