Introduce SegmentBlueprint
SegmentBlueprint is intended to replace VerticalWindow.
Essentially it is the same class, but rewritten to suite new objects
model.

VerticalWindow was initially intended as a cardboard window which
travels along the surface and shows what shape data takes at this point.
With time its purpose changed slightly from the original one and no
longer represents the idea the best way it could.

SegmentBlueprint also describes a shape of any abstract data segment
over the data.

Implementation is done in the header as definitions are short and
supposed to be called a lot.