Introduce SurfaceBoundedSubVolume
Class is intended to replace Horizon and combine in itself some
previously loose functionality, like cubic_makima, segment offsets and
resampling logic.

SurfaceBoundedSubVolume is supposed to be a set of data segments over
the horizontal plane.
Data is stored continuously, which makes it suitable to be filled by
openvds functions.
The class is not supposed to be used for resampling as we simply do not
need a big contiguous chunk of memory for attributes calculations.