Add position of minimum value attribute
Added a new attribute 'min_at' that will return the position of the
minimum value in sample axes for both along the surface and between the
surfaces.

Note: If there are multiple instances of the same value, it will return
the position of the first occurrence of that value in sample axes.