Add position of maximum value attribute
Added a new attribute 'max_at' that will return the position of the
maximum value in sample axes for both along the surface and between the
surfaces.

Note: If there are multiple instances of the same value, it will return
the position of the first occurrence of that value in sample axes.