Add position of maximum absolute value attribute
Added a new attribute 'maxabs_at' that will return the position of the
maximum absolute value in sample axes for both along the surface and
between the surfaces.

Note: If there are multiple instances of the same value, it will return
the position of the first occurrence of that value in sample axes.