Remove unnecessary e2e testcase
In our e2e test case, we  download the latest version of the blob
library every time we run new container. However, the recent release of
azure-storage-blob 12.19.0 has made it mandatory to include the
permission argument in order to generate a SAS token.

Also, Since Azure is responsible for ensuring the validity of SAS tokens
and granting appropriate access permissions. Testing that with empty
permission is  unnecessary for us to test. Therefore, removing that
particular test case.

[ref](https://github.com/Azure/azure-sdk-for-python/commit/86759fffbae2c93771607aacd2e6e5697b5b0bda)