Make CMake search for Boost libraries in C++ tests
Building the tests were failing on systems where Boost is not in the
default search path for include directories, e.g., if Boost was
installed with Homebrew.

The change also makes the dependency on liboost more explicit as CMake
will print an error message if CMake does not find Boost during the
configuration stage.