Use jemalloc for memory allocation
During recent investigations we found out that memory allocation has a
major impact on performance of multi-threading application that creates
many objects on heap. That is our case.

Effects of malloc are more visible for big files.
Using better malloc can improve code performance from 10% to 75%,
depending on threading.

We chose jemalloc as it is built on alpine 3.17 and gives best
performance improvement. We use it with LD_PRELOAD by supplying library
during runtime.