Drop 2 margin samples boundary condition
It has been agreed that we want to drop the requirement of reserving two
samples before the boundary for attributes calculations.
As long as user's request falls inside vertical bounds, we want it to
succeed.

When less than 2 samples are reserved for calculations near the border,
calculations will become less precise. However, as this data already is
near the border, users should expect the data to be of questionable
quality anyway.