Combine fetchSubvolume and calculateAttributes
FetchSubvolume and calculateAttributes were nice steps separations.
However it is not the most efficient way to perform calculations.

When some of the goroutines are finished fetching subvolume and there is
nothing for them to pick up anymore, they become idle, when instead they
could already begin on attributes calculations.

As benchmarking shows, fetching and calculating together instead of in
separate steps can speed up attributes call almost by full time spent in
calculateAttributes.

Note for the future: it can happen that if we move to multiple
datahandles, performance gain from that change will consume the current
speedup, but for now this change seems reasonable.

Additionaly, this change allows to get rid of subvolume as a contineous
data storing space, if the need ever arises. We wouldn't need to store
whole subvolume data and will be able to reuse the memory after
computations on chunk are complete.
At the moment though subvolume structure seems nicer to have.