Use different approach to calculate grid metadata
Original idea was to expose grid metadata to the user and thus bypass
these calculations in tests altogether. However it was decided against
directly exposing it as it would be confusing for the end user because
of the inconsistencies in the domain.

So calculations are changed instead to
- get rid of the openvds coordinate transformer dependency
- illustrate the way user might get grid information from currently
exposed data