Fix size calculation for RawSegment
Raw segments do not manage their own data, they get it from outside.
That data might be, for example, empty, when we are dealing with
segments created in the place of fill values.

This case wasn't considered before and thus size was calculated from the
blueprint, not from the actual data.

Resampled segments, on the other hand, manage their own data, so they
are required to have the blueprint size and are not subject to this
issue.