Forgo openvds dependency in Axis constructor
Currently Axis is a simple wrapper over AxisDescriptor. AxisDescriptor
is responsible not only for nsamples, min and max, but also for name and
unit, which in openvds world are char*. It means that our Axis
functionality depends on some external memory we do not control.

With introduction of double metadata this becomes a problem if we try to
minimize number of dependencies.
So idea is to change Axis from being a wrapper over AxisDescriptor to
its own structure that just uses AxisDescriptor class for calculations.