Directly validate direction uniqueness
Currently that becomes a duplicated check as for now it is covered by
axes creation (given axis name, find its dimension) as each of the names
is requested just once.
But as logic is planned to be inverted (given dimension, find axis
name), we want to have this check explicitly.