Add trusted proxies option
Add an option to set a list of trusted proxies when starting the server.
This option will determine which IP is considered to be the source IP of
a request when headers that contain alternative IP adresses are set (if
the proxy is trusted the IP the request is forwarded for will be
considered rthe source).

As far as we know the only part of our system
that currently will be impacted by this is the logging.