Move slice request handeling to separate file
The code was spread between two files: request.go and enpoint.go. The
filenames did not clearly convey the expected separation between these
files making it challenging to navigate the code.

This is the first in a series of commits attempting to clean up this
code by splitting it up in separate files for each of the endpoints as
well as using more descriptive and standard naming conventions.