Add cube intersection annotation-based checks
Currently check method parameters are based on index values, but not on
cube index values - on index values corresponding to custom arrays
defined in every file (inline_array, xline_array, sample_array).

As experience shown, this is very confusing and time-consuming both when
old tests need to be re-checked/refactored as well as when new tests are
added.

Not only one has to remember that values relate to some array, one has
to calculate which annotation these values correspond to.

Also test readability suffers as values written down in tests do not
obviously correspond to anything.

Hence the idea to change values to annotation coordinate system. As all
the cubes have a help hint which annotated lines they contain, it is
fast and easy to calculate what intersection of two each cubes should
contain.