Tolerate any sample as zero sample in subvolume
Support for negative values made us reconsider meaning of zero sample.
Originally it was the smallest value that is less than stepsize.

With recent refactorings it seems this meaning doesn't matter any
longer.
Zero sample can be any sample one considers to be "zero" for index
calculations or for calculations of several things that suppose to have
the same anchor.

So the meaning is relaxed.