Fix wrong fence hash on FillValue field
Apparently our fence hashing mechanism worked wrong.
FillValue is an optional value of type *float32.
When it is not supplied, it gets assigned the value of nil pointer.

However, gob serialization library seems to deal badly with nil pointers
and in just stores the default value of pointed to type, in this case 0.

So if FillValue points to 0, it becomes indistinguishable from nil
pointer.

So if user sends a request with FillValue=0, result is cached under
common hash. And then result of the request with FillValue not supplied
returns that cached result.
So we returned user incorrect data.

Now requests with and without FillValue supplied have different hashes.

As a side effect, now there could be two different cache entries with
same results for good calls (one without FillValue supplied and one with
it set to 0). This data duplication is acceptable as it solves more
severe problem and will unlikely appear very often anyway.