Close openvds resources
We are not closing opened openvds resources. Again.
We already had this issue a couple years back, but we managed to forget
closing once more during refactoring a while ago.

It looks like this memory leak might be present on our monitoring
dashboard: 2 weeks in we are using several GB more than just after
container restart and more than assigned max cache memory.
(At least this seems to be the most probable cause).

This problem led to resurrection of idea of memory testing.

Note that not closing openvds resources doesn't seem to be the proper
memory leak: even if we do not have pointer to datahandle any longer,
tools still don't report memory leak.
Possible cause is that openvds registers something in the global scope
or created threads have something to do with that.