Don't store bottom margin in subvolume
The value was put there intentionally to be symmetrical with
top_margin.
However the need for it doesn't seem to come and it actually appears
confusing once one forgets about it.
It is better to remove it now and then recreate if needed based on
symmetry with top margin.

We don't actually need it as other information (top margin, segment
size, top boundary) are enough to define all the data we need.