Prevent failure on insufficient samples number
There is a user case where provided two horizons almost touch each other
near the survey border. Users are not interested in this interval, but
they are interested in query not failing due to that.

Hence path of least resistance is implemented - if we do not have enough
samples to satisfy makima algorithms conditions, take lacking samples
from the other side of the data.
Continuity is unbroken and makima will still interpolate data in the
region between two last samples.