Rename golang package to oneseismic-api
Package is published to pkg.go.dev, but as it is a side effect of golang
infrastructure, we will just ignore it.
It is very unlikely that anyone besides us would be using our solution
as a library.
Also, we are not overriding the package, it would just be published
under new name.