Make number of azure threads configurable
In 3.4.4 open-vds introduced a way to choose number of threads used to
download data with Azure.

From experiments it might be better to decide on number of threads per
file, but
- number of threads is set through openvds environment variable
OPENVDS_AZURE_IOMANAGER_THREAD_COUNT, so one can't really define this
value per request
- there is some fluctuation in the results
- there doesn't seem to be extremely significant difference when
larger-than-desired number of threads is chosen
- whatever formula we come up with would be heuristics anyway

So for simplicity variable is defined for the whole server instance.
Experiments have shown that best value lies somewhere between ncores and
ncores x 4.
It seems like with x4 cores we get best speed, but random slower times
start to pop up.