Monitor openvds with snyk
By default openvds statically builds all dependencies. Some we override
to be used dynamically, for some we rely on openvds version.

snyk --unmanaged will check build directory for any vulnerabilities
known to it. While we might not be affected by all of them, it would be
good to have an overview.