Remove medtime and maxtime from suits
Due to severe fluctuations in the execution time, these ones are not
used and basically are ignored. So they can be removed for now, at least
until we have thresholds provided by the PO.

They are still left in manual execution though, as this one is used for
research purpose. I doubt though they would be highly useful there as
well, and are likely to be removed completely at the first occasion in
the future.